Renaming existing random-beacon.ts to fake-random-beacon.ts
The rename was done because here we fake the Random Beacon contract and
the real contract will be used in other tests.
